requirements.txt
requests>=2.31.0
pydantic>=2.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
APScheduler>=3.10.4
